Moved on & off functions to own, top level, utils file now they are no longer used solely for SecuredFields
