+++ packages/lib/src/components/internal/SecuredFields/lib/utilities/dom.ts
@@ -26,24 +26,6 @@ const getAttribute = (node, attribute) => {
     return node.getAttribute(attribute);
 };
 
-const on = (node, event, callback, useCapture?) => {
-    if (typeof node.addEventListener === 'function') {
-        node.addEventListener(event, callback, useCapture);
-        return;
-    }
-
-    throw new Error(`: Unable to bind ${event}-event`);
-};
-
-const off = (node, event, callback, useCapture?) => {
-    if (typeof node.addEventListener === 'function') {
-        node.removeEventListener(event, callback, useCapture);
-        return;
-    }
-
-    throw new Error(`: Unable to unbind ${event}-event`);
-};
-
 /**
  * @internal
  * Removes all children elements of the given node.
@@ -56,4 +38,4 @@ const removeAllChildren = node => {
     }
 };
 
-export { getAttribute, on, off, select, selectOne, removeAllChildren };
+export { getAttribute, select, selectOne, removeAllChildren };
